[Bug] Sheet priorities weren't deleting correctly

+++ resources/js/state/sheet/actions/deleteSheetPriority.tsx
@@ -3,9 +3,13 @@
 //-----------------------------------------------------------------------------
 import { mutation } from '@/api'
 
+import { SHEET_PRIORITY_IS_IN_USE } from '@/state/messenger/messages'
+
 import { IAppState } from '@/state'
 import { IThunkAction, IThunkDispatch } from '@/state/types'
-import { ISheet, ISheetPriority, ISheetCellPriority } from '@/state/sheet/types'
+import { ISheet, ISheetPriority } from '@/state/sheet/types'
+
+import { createMessengerMessage } from '@/state/messenger/actions' 
 import { 
   setAllSheetPriorities,
   updateSheet
@@ -24,25 +28,39 @@ export const deleteSheetPriority = (sheetId: ISheet['id'], sheetPriorityId: IShe
       allSheetPriorities
     } = getState().sheet
 
-    const { [sheetPriorityId]: deletedSheetPriority, ...nextAllSheetPriorities } = allSheetPriorities
-
-    const nextSheetPriorities = sheet.priorities.filter(priorityId => priorityId !== sheetPriorityId)
-    
-    const nextSheetCellPriorities: { [cellId: string]: ISheetCellPriority } = {}
+    // Make sure there are no cells using the priority we're trying to delete
+    let isSheetPriorityInUse = false
     Object.keys(sheet.cellPriorities).forEach(cellId => {
       const sheetCellPriority = sheet.cellPriorities[cellId]
-      if(sheetCellPriority.id !== sheetPriorityId) {
-        nextSheetCellPriorities[cellId] = sheetCellPriority
+      console.log(sheetCellPriority.priorityId, sheetPriorityId)
+      if(sheetCellPriority) {
+        if(sheetCellPriority.priorityId === sheetPriorityId) {
+          isSheetPriorityInUse = true
+        }
       }
     })
 
-    dispatch(setAllSheetPriorities(nextAllSheetPriorities))
+    // If there aren't any cells using the priority
+    if(!isSheetPriorityInUse) {
 
-    dispatch(updateSheet(sheetId, {
-      priorities: nextSheetPriorities,
-      cellPriorities: nextSheetCellPriorities
-    }, true))
-
-    mutation.deleteSheetPriority(sheetPriorityId)
+      // Remove the priority from all sheet priorities
+      const { [sheetPriorityId]: deletedSheetPriority, ...nextAllSheetPriorities } = allSheetPriorities
+  
+      // Filter the sheet priorities, removing the priority we're deleting
+      const nextSheetPriorities = sheet.priorities.filter(priorityId => priorityId !== sheetPriorityId)
+  
+      // Update the store
+      dispatch(setAllSheetPriorities(nextAllSheetPriorities))
+      dispatch(updateSheet(sheetId, {
+        priorities: nextSheetPriorities
+      }, true))
+  
+      // Delete from the database
+      mutation.deleteSheetPriority(sheetPriorityId)
+    }
+    // If there are cells using the priority, display an error message
+    else {
+      dispatch(createMessengerMessage(SHEET_PRIORITY_IS_IN_USE))
+    }
   }
 }
